# index: optimize go GC performance