What does this PR change? What problem does it solve?
-----------------------------------------------------

The current index implementation using a custom indexmap is not particular garbage-collection friendly. As each indexEntry and the bucket list contain pointers, this requires the garbage collector to chase pointers to random locations throughout the index.

This PR moves the indexEntries into an array and only adds the array index instead of a pointer. Thereby, the indexEntries and bucket list no longer contain pointers and are not scanned by the garbage collector. The array containing all indexEntries is implemented as https://en.wikipedia.org/wiki/Hashed_array_tree to minimize the memory overhead.

The top-layer of the hashed array tree (HAT) also has a size of O(sqrt(n)), which should make it cache efficient. The top-layer should be small enough to easily fit into the CPU cache and thus only adds little overhead compared to directly accessing an index entry via a pointer.

As a bonus, the indexmap.Each/Grow functions are now able to iterate over the indexEntries array instead of following random pointers. This significantly speeds up both functions.

## Results

The indirection through the HAT increases the index lookup duration by 20-40% (the average is probably between 20-30% but the results are pretty noisy as a result of the varying map entry layout due to the maphash randomization), although the MasterIndex based tests show numbers between 0 to 35% overhead.

Iterating over the index is now up to 8 times faster, which also translates to slightly faster index encoding.
GC is now nearly for free.

grow() is also pretty interesting, the performance changes are likely a result of a different order of the chained indexEntries.


Commits:

- fast-grow `index: implement indexmap.grow() without random access`
- fast-each `index: let indexmap.Each iterate in allocation order`
- hat-opt `index: remove redundant storage of indexmap size`
- hat-inline `index: Allow inlining of HAT`
- hat-base `index: store indexEntries in hashed array tree`
- remove-pointers `index: Remove pointers from within indexentrys`
- base - `index: add garbage collection benchmark`

<details>

<summary>benchstat base remove-pointers</summary>

```
name                                              old time/op    new time/op    delta
IndexMapHash-12                                     1.25µs ± 3%    1.27µs ± 1%    +1.62%  (p=0.000 n=15+15)
DecodeIndex-12                                       2.38s ± 4%     2.35s ± 1%      ~     (p=0.367 n=15+15)
DecodeIndexParallel-12                               2.31s ± 5%     2.38s ± 1%    +2.70%  (p=0.041 n=15+12)
EncodeIndex/100-12                                   618µs ± 3%     621µs ± 3%      ~     (p=0.233 n=15+15)
EncodeIndex/1000-12                                 6.05ms ± 3%    6.11ms ± 2%      ~     (p=0.074 n=15+15)
EncodeIndex/10000-12                                64.5ms ± 2%    63.9ms ± 2%      ~     (p=0.134 n=15+14)
IndexHasUnknown-12                                  25.2ns ± 9%    26.7ns ±16%      ~     (p=0.083 n=14+15)
IndexHasKnown-12                                    26.8ns ± 7%    27.7ns ±11%    +3.35%  (p=0.030 n=13+15)
IndexAlloc-12                                        301ms ± 1%     338ms ± 2%   +12.34%  (p=0.000 n=15+15)
IndexAllocParallel-12                                107ms ± 3%      92ms ± 2%   -13.80%  (p=0.000 n=15+14)
MasterIndexAlloc-12                                  133ms ± 2%     197ms ± 3%   +48.36%  (p=0.000 n=15+15)
MasterIndexLookupSingleIndex-12                     60.3ns ±10%    61.3ns ± 9%      ~     (p=0.305 n=15+15)
MasterIndexLookupMultipleIndex-12                   60.4ns ± 8%    54.9ns ± 6%    -9.22%  (p=0.000 n=15+14)
MasterIndexLookupSingleIndexUnknown-12              31.9ns ±11%    33.9ns ±12%    +6.27%  (p=0.015 n=15+15)
MasterIndexLookupMultipleIndexUnknown-12            30.4ns ± 9%    31.3ns ± 6%      ~     (p=0.083 n=15+15)
MasterIndexLookupParallel/known,indices=25-12        163ns ± 9%     165ns ±12%      ~     (p=0.723 n=14+15)
MasterIndexLookupParallel/unknown,indices=25-12      129ns ±13%     133ns ±19%      ~     (p=0.330 n=15+15)
MasterIndexLookupParallel/known,indices=50-12        144ns ±14%     166ns ±13%   +15.12%  (p=0.000 n=15+15)
MasterIndexLookupParallel/unknown,indices=50-12      128ns ±11%     133ns ±14%      ~     (p=0.075 n=14+15)
MasterIndexLookupParallel/known,indices=100-12       144ns ±15%     160ns ±18%   +10.89%  (p=0.004 n=15+15)
MasterIndexLookupParallel/unknown,indices=100-12     130ns ± 5%     130ns ±13%      ~     (p=0.892 n=13+15)
MasterIndexLookupBlobSize-12                        30.3ns ± 6%    31.2ns ± 7%      ~     (p=0.087 n=15+15)
MasterIndexEach-12                                   734ms ± 1%     669ms ± 3%    -8.97%  (p=0.000 n=15+15)
MasterIndexGC-12                                     116ms ±36%       0ms ± 2%   -99.74%  (p=0.000 n=15+12)

name                                              old speed      new speed      delta
IndexMapHash-12                                   3.29GB/s ± 3%  3.23GB/s ± 1%    -1.60%  (p=0.000 n=15+15)

name                                              old alloc/op   new alloc/op   delta
IndexMapHash-12                                      0.00B          0.00B           ~     (all equal)
EncodeIndex/100-12                                   263kB ± 0%     263kB ± 0%      ~     (p=0.926 n=15+15)
EncodeIndex/1000-12                                 2.61MB ± 0%    2.61MB ± 0%      ~     (p=0.516 n=15+12)
EncodeIndex/10000-12                                25.7MB ± 2%    25.7MB ± 5%      ~     (p=0.775 n=15+15)
IndexAlloc-12                                        451MB ± 0%     954MB ± 0%  +111.49%  (p=0.000 n=13+15)
IndexAllocParallel-12                                451MB ± 0%     954MB ± 0%  +111.46%  (p=0.000 n=11+14)
MasterIndexAlloc-12                                  162MB ± 0%     347MB ± 0%  +114.17%  (p=0.000 n=15+13)

name                                              old allocs/op  new allocs/op  delta
IndexMapHash-12                                       0.00           0.00           ~     (all equal)
EncodeIndex/100-12                                   2.39k ± 0%     2.39k ± 0%      ~     (all equal)
EncodeIndex/1000-12                                  22.9k ± 0%     22.9k ± 0%      ~     (p=0.061 n=15+13)
EncodeIndex/10000-12                                  229k ± 0%      229k ± 0%      ~     (p=0.690 n=15+15)
IndexAlloc-12                                        1.53M ± 0%     1.10M ± 0%   -28.25%  (p=0.000 n=14+15)
IndexAllocParallel-12                                1.53M ± 0%     1.10M ± 0%   -28.25%  (p=0.000 n=11+15)
MasterIndexAlloc-12                                   585k ± 0%      435k ± 0%   -25.67%  (p=0.000 n=15+13)
```

</details>

<details>

<summary>benchstat base hat-base</summary>

```
name                                              old time/op    new time/op    delta
IndexMapHash-12                                     1.25µs ± 3%    1.22µs ± 2%   -1.86%  (p=0.000 n=15+14)
DecodeIndex-12                                       2.38s ± 4%     2.28s ± 1%   -4.14%  (p=0.000 n=15+14)
DecodeIndexParallel-12                               2.31s ± 5%     2.35s ± 4%     ~     (p=0.056 n=15+15)
EncodeIndex/100-12                                   618µs ± 3%     623µs ± 3%     ~     (p=0.074 n=15+15)
EncodeIndex/1000-12                                 6.05ms ± 3%    6.06ms ± 1%     ~     (p=0.591 n=15+14)
EncodeIndex/10000-12                                64.5ms ± 2%    64.6ms ± 1%     ~     (p=0.186 n=15+14)
IndexHasUnknown-12                                  25.2ns ± 9%    35.0ns ±29%  +39.01%  (p=0.000 n=14+15)
IndexHasKnown-12                                    26.8ns ± 7%    32.5ns ±19%  +21.18%  (p=0.000 n=13+15)
IndexAlloc-12                                        301ms ± 1%     288ms ± 2%   -4.25%  (p=0.000 n=15+15)
IndexAllocParallel-12                                107ms ± 3%      70ms ± 1%  -33.90%  (p=0.000 n=15+14)
MasterIndexAlloc-12                                  133ms ± 2%     147ms ± 3%  +10.52%  (p=0.000 n=15+15)
MasterIndexLookupSingleIndex-12                     60.3ns ±10%    67.2ns ±17%  +11.33%  (p=0.000 n=15+15)
MasterIndexLookupMultipleIndex-12                   60.4ns ± 8%    60.4ns ±11%     ~     (p=0.645 n=15+15)
MasterIndexLookupSingleIndexUnknown-12              31.9ns ±11%    40.2ns ±21%  +25.83%  (p=0.000 n=15+13)
MasterIndexLookupMultipleIndexUnknown-12            30.4ns ± 9%    36.9ns ±25%  +21.36%  (p=0.000 n=15+14)
MasterIndexLookupParallel/known,indices=25-12        163ns ± 9%     178ns ±18%   +8.89%  (p=0.003 n=14+15)
MasterIndexLookupParallel/unknown,indices=25-12      129ns ±13%     136ns ±19%     ~     (p=0.083 n=15+15)
MasterIndexLookupParallel/known,indices=50-12        144ns ±14%     178ns ±15%  +23.18%  (p=0.000 n=15+15)
MasterIndexLookupParallel/unknown,indices=50-12      128ns ±11%     140ns ± 7%   +9.33%  (p=0.000 n=14+13)
MasterIndexLookupParallel/known,indices=100-12       144ns ±15%     174ns ±16%  +20.81%  (p=0.000 n=15+15)
MasterIndexLookupParallel/unknown,indices=100-12     130ns ± 5%     140ns ±15%   +7.60%  (p=0.011 n=13+15)
MasterIndexLookupBlobSize-12                        30.3ns ± 6%    35.3ns ±16%  +16.50%  (p=0.000 n=15+14)
MasterIndexEach-12                                   734ms ± 1%     706ms ± 4%   -3.85%  (p=0.000 n=15+15)
MasterIndexGC-12                                     116ms ±36%       0ms ±25%  -99.73%  (p=0.000 n=15+13)

name                                              old speed      new speed      delta
IndexMapHash-12                                   3.29GB/s ± 3%  3.35GB/s ± 2%   +1.89%  (p=0.000 n=15+14)

name                                              old alloc/op   new alloc/op   delta
IndexMapHash-12                                      0.00B          0.00B          ~     (all equal)
EncodeIndex/100-12                                   263kB ± 0%     263kB ± 0%   +0.02%  (p=0.041 n=15+15)
EncodeIndex/1000-12                                 2.61MB ± 0%    2.61MB ± 1%     ~     (p=0.690 n=15+15)
EncodeIndex/10000-12                                25.7MB ± 2%    25.4MB ± 0%     ~     (p=0.129 n=15+12)
IndexAlloc-12                                        451MB ± 0%     496MB ± 0%   +9.95%  (p=0.000 n=13+15)
IndexAllocParallel-12                                451MB ± 0%     496MB ± 0%   +9.95%  (p=0.000 n=11+15)
MasterIndexAlloc-12                                  162MB ± 0%     213MB ± 0%  +31.58%  (p=0.000 n=15+13)

name                                              old allocs/op  new allocs/op  delta
IndexMapHash-12                                       0.00           0.00          ~     (all equal)
EncodeIndex/100-12                                   2.39k ± 0%     2.39k ± 0%     ~     (all equal)
EncodeIndex/1000-12                                  22.9k ± 0%     22.9k ± 0%     ~     (p=0.074 n=15+12)
EncodeIndex/10000-12                                  229k ± 0%      229k ± 0%     ~     (p=0.927 n=15+15)
IndexAlloc-12                                        1.53M ± 0%     1.10M ± 0%  -28.13%  (p=0.000 n=14+15)
IndexAllocParallel-12                                1.53M ± 0%     1.10M ± 0%  -28.13%  (p=0.000 n=11+15)
MasterIndexAlloc-12                                   585k ± 0%      507k ± 0%  -13.33%  (p=0.000 n=15+14)
```

</details>

<details>

<summary>benchstat base hat-inline</summary>

```
name                                              old time/op    new time/op    delta
IndexMapHash-12                                     1.25µs ± 3%    1.25µs ± 2%     ~     (p=0.751 n=15+15)
DecodeIndex-12                                       2.38s ± 4%     2.28s ± 1%   -4.17%  (p=0.000 n=15+14)
DecodeIndexParallel-12                               2.31s ± 5%     2.33s ± 5%     ~     (p=0.267 n=15+15)
EncodeIndex/100-12                                   618µs ± 3%     624µs ± 5%     ~     (p=0.137 n=15+15)
EncodeIndex/1000-12                                 6.05ms ± 3%    6.04ms ± 1%     ~     (p=0.983 n=15+14)
EncodeIndex/10000-12                                64.5ms ± 2%    65.1ms ± 3%   +0.90%  (p=0.033 n=15+15)
IndexHasUnknown-12                                  25.2ns ± 9%    33.4ns ±39%  +32.65%  (p=0.000 n=14+15)
IndexHasKnown-12                                    26.8ns ± 7%    34.4ns ±36%  +28.24%  (p=0.000 n=13+15)
IndexAlloc-12                                        301ms ± 1%     286ms ± 2%   -4.81%  (p=0.000 n=15+15)
IndexAllocParallel-12                                107ms ± 3%      70ms ± 2%  -34.48%  (p=0.000 n=15+15)
MasterIndexAlloc-12                                  133ms ± 2%     146ms ± 3%  +10.06%  (p=0.000 n=15+15)
MasterIndexLookupSingleIndex-12                     60.3ns ±10%    67.5ns ±16%  +11.92%  (p=0.001 n=15+14)
MasterIndexLookupMultipleIndex-12                   60.4ns ± 8%    60.9ns ±14%     ~     (p=0.775 n=15+15)
MasterIndexLookupSingleIndexUnknown-12              31.9ns ±11%    42.6ns ±37%  +33.55%  (p=0.000 n=15+15)
MasterIndexLookupMultipleIndexUnknown-12            30.4ns ± 9%    34.8ns ±19%  +14.52%  (p=0.000 n=15+15)
MasterIndexLookupParallel/known,indices=25-12        163ns ± 9%     173ns ±15%   +6.01%  (p=0.010 n=14+15)
MasterIndexLookupParallel/unknown,indices=25-12      129ns ±13%     140ns ±18%   +8.97%  (p=0.003 n=15+15)
MasterIndexLookupParallel/known,indices=50-12        144ns ±14%     171ns ±16%  +19.01%  (p=0.000 n=15+15)
MasterIndexLookupParallel/unknown,indices=50-12      128ns ±11%     148ns ±13%  +15.12%  (p=0.000 n=14+14)
MasterIndexLookupParallel/known,indices=100-12       144ns ±15%     170ns ±16%  +17.83%  (p=0.000 n=15+15)
MasterIndexLookupParallel/unknown,indices=100-12     130ns ± 5%     142ns ±21%   +9.01%  (p=0.004 n=13+15)
MasterIndexLookupBlobSize-12                        30.3ns ± 6%    36.0ns ±18%  +18.87%  (p=0.000 n=15+15)
MasterIndexEach-12                                   734ms ± 1%     710ms ± 5%   -3.33%  (p=0.000 n=15+15)
MasterIndexGC-12                                     116ms ±36%       0ms ±28%  -99.64%  (p=0.000 n=15+15)

name                                              old speed      new speed      delta
IndexMapHash-12                                   3.29GB/s ± 3%  3.28GB/s ± 2%     ~     (p=0.784 n=15+15)

name                                              old alloc/op   new alloc/op   delta
IndexMapHash-12                                      0.00B          0.00B          ~     (all equal)
EncodeIndex/100-12                                   263kB ± 0%     263kB ± 0%     ~     (p=0.862 n=15+15)
EncodeIndex/1000-12                                 2.61MB ± 0%    2.61MB ± 1%     ~     (p=0.624 n=15+15)
EncodeIndex/10000-12                                25.7MB ± 2%    25.8MB ± 5%     ~     (p=0.683 n=15+15)
IndexAlloc-12                                        451MB ± 0%     496MB ± 0%   +9.95%  (p=0.000 n=13+15)
IndexAllocParallel-12                                451MB ± 0%     496MB ± 0%   +9.95%  (p=0.000 n=11+15)
MasterIndexAlloc-12                                  162MB ± 0%     213MB ± 0%  +31.58%  (p=0.000 n=15+11)

name                                              old allocs/op  new allocs/op  delta
IndexMapHash-12                                       0.00           0.00          ~     (all equal)
EncodeIndex/100-12                                   2.39k ± 0%     2.39k ± 0%     ~     (all equal)
EncodeIndex/1000-12                                  22.9k ± 0%     22.9k ± 0%     ~     (p=0.061 n=15+13)
EncodeIndex/10000-12                                  229k ± 0%      229k ± 0%     ~     (p=0.719 n=15+15)
IndexAlloc-12                                        1.53M ± 0%     1.10M ± 0%  -28.13%  (p=0.000 n=14+15)
IndexAllocParallel-12                                1.53M ± 0%     1.10M ± 0%  -28.13%  (p=0.000 n=11+15)
MasterIndexAlloc-12                                   585k ± 0%      507k ± 0%  -13.33%  (p=0.000 n=15+11)
```

</details>


<details>

<summary>benchstat base hat-opt</summary>

```
name                                              old time/op    new time/op    delta
IndexMapHash-12                                     1.25µs ± 3%    1.24µs ± 1%     ~     (p=0.082 n=15+14)
DecodeIndex-12                                       2.38s ± 4%     2.28s ± 1%   -4.04%  (p=0.000 n=15+15)
DecodeIndexParallel-12                               2.31s ± 5%     2.34s ± 5%     ~     (p=0.148 n=15+15)
EncodeIndex/100-12                                   618µs ± 3%     627µs ± 2%   +1.55%  (p=0.002 n=15+15)
EncodeIndex/1000-12                                 6.05ms ± 3%    6.05ms ± 2%     ~     (p=0.870 n=15+15)
EncodeIndex/10000-12                                64.5ms ± 2%    64.4ms ± 2%     ~     (p=0.496 n=15+13)
IndexHasUnknown-12                                  25.2ns ± 9%    30.6ns ±13%  +21.68%  (p=0.000 n=14+14)
IndexHasKnown-12                                    26.8ns ± 7%    33.4ns ±23%  +24.61%  (p=0.000 n=13+15)
IndexAlloc-12                                        301ms ± 1%     286ms ± 2%   -4.88%  (p=0.000 n=15+14)
IndexAllocParallel-12                                107ms ± 3%      70ms ± 3%  -34.63%  (p=0.000 n=15+15)
MasterIndexAlloc-12                                  133ms ± 2%     144ms ± 2%   +8.70%  (p=0.000 n=15+15)
MasterIndexLookupSingleIndex-12                     60.3ns ±10%    66.2ns ±16%   +9.69%  (p=0.004 n=15+15)
MasterIndexLookupMultipleIndex-12                   60.4ns ± 8%    60.6ns ±13%     ~     (p=0.992 n=15+15)
MasterIndexLookupSingleIndexUnknown-12              31.9ns ±11%    40.9ns ±21%  +28.16%  (p=0.000 n=15+15)
MasterIndexLookupMultipleIndexUnknown-12            30.4ns ± 9%    36.6ns ±27%  +20.32%  (p=0.000 n=15+15)
MasterIndexLookupParallel/known,indices=25-12        163ns ± 9%     173ns ±19%   +5.90%  (p=0.014 n=14+15)
MasterIndexLookupParallel/unknown,indices=25-12      129ns ±13%     143ns ±13%  +11.12%  (p=0.000 n=15+14)
MasterIndexLookupParallel/known,indices=50-12        144ns ±14%     176ns ±14%  +22.23%  (p=0.000 n=15+14)
MasterIndexLookupParallel/unknown,indices=50-12      128ns ±11%     132ns ±19%     ~     (p=0.598 n=14+15)
MasterIndexLookupParallel/known,indices=100-12       144ns ±15%     176ns ±14%  +21.67%  (p=0.000 n=15+15)
MasterIndexLookupParallel/unknown,indices=100-12     130ns ± 5%     138ns ±17%     ~     (p=0.060 n=13+15)
MasterIndexLookupBlobSize-12                        30.3ns ± 6%    37.5ns ±22%  +23.92%  (p=0.000 n=15+15)
MasterIndexEach-12                                   734ms ± 1%     696ms ± 6%   -5.26%  (p=0.000 n=15+15)
MasterIndexGC-12                                     116ms ±36%       0ms ±25%  -99.68%  (p=0.000 n=15+15)

name                                              old speed      new speed      delta
IndexMapHash-12                                   3.29GB/s ± 3%  3.31GB/s ± 1%     ~     (p=0.085 n=15+14)

name                                              old alloc/op   new alloc/op   delta
IndexMapHash-12                                      0.00B          0.00B          ~     (all equal)
EncodeIndex/100-12                                   263kB ± 0%     263kB ± 0%     ~     (p=0.158 n=15+15)
EncodeIndex/1000-12                                 2.61MB ± 0%    2.61MB ± 1%     ~     (p=0.461 n=15+15)
EncodeIndex/10000-12                                25.7MB ± 2%    25.4MB ± 0%   -1.16%  (p=0.029 n=15+13)
IndexAlloc-12                                        451MB ± 0%     496MB ± 0%   +9.95%  (p=0.000 n=13+15)
IndexAllocParallel-12                                451MB ± 0%     496MB ± 0%   +9.95%  (p=0.000 n=11+15)
MasterIndexAlloc-12                                  162MB ± 0%     213MB ± 0%  +31.58%  (p=0.000 n=15+11)

…