@MichaelEischer 2023-06-02
To complement the above measurements, here are the number of a full system backup dry-run with 140GB in 1 Million files:

master - 106,02s user 12,96s system 265% cpu 44,855 total
pointerless-index - 87,52s user 12,64s system 232% cpu 43,112 total

That is, with this PR the backup is faster and requires 17% less CPU! For larger repositories, the improvement is likely even larger.